REFACTOR: pages폴더에 각 페이지 컴포넌트 이동 -기존의 components에 몰아서 저장하던 방식에서 pages에 페이지 컴포넌트를 보관하는 방식으로 수정 -component폴더에는 각 페이지별로 폴더링해서 파일관리 -페이지 컴포넌트 이동으로 인한 경로 변경사항 수정
