+++ src/App.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { BrowserRouter } from "react-router-dom";
 import { RecoilRoot } from "recoil";
-import Router from "./components/Router";
+import Router from "./pages/Router";
 import MenuBar from "./components/MenuBar";
 import StatusBar from "./components/StatusBar";
 import * as S from "./styles/App.styled";
