+++ src/pages/Router.js
@@ -1,17 +1,17 @@
 import { Route, Routes } from "react-router-dom";
-import ProfileListView from "./UserProfile/ProfileListView";
-import ChattingList from "./ChattingList/ChattingListView";
-import ChattingView from "./Chatting/MessageChattingView";
+import UserProfile from "./UserProfile";
+import ChattingList from "./ChattingList";
+import Chatting from "./Chatting";
 import MoreView from "./MoreView";
 
 const Router = () => {
   return (
     <>
       <Routes>
-        <Route path="/profile" element={<ProfileListView />} />
+        <Route path="/profile" element={<UserProfile />} />
         <Route path="/chatting-list" element={<ChattingList />} />
-        <Route path="/chat-with/:userID" element={<ChattingView />} />
         <Route path="/more" element={<MoreView />} />
+        <Route path="/chat-with/:userID" element={<Chatting />} />
       </Routes>
     </>
   );
